+++ HISTORY.rst
@@ -2,10 +2,21 @@
 History
 =======
 
-0.7.2 (unreleased)
+0.8.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+Improvements
+
+- enable search on several resource types (_type search param) [Jasopaum]
+
+- Issue #8 `Add search support for without any params or query string if context has resource type <https://github.com/nazrulworld/fhirpath/issues/8>`_ [nazrulworld]
+
+- Issue #9 `multiple negative not working <https://github.com/nazrulworld/fhirpath/issues/9>`_ [nazrulworld]
+
+Breaking
+
+- ``fhirpath.search.SearchContext.resource_name`` has been changed ``fhirpath.search.SearchContext.resource_type`` and
+  now datatype is List instead of string. Please check your API. [Jasopaum]
 
 
 0.7.1 (2020-08-07)